apps/sandbox/pyproject.toml
[tool.poetry]
name = "shadow-agent-sandbox"
version = "0.1.0"
description = "Sandbox template for Shadow coding agent with git and Node.js"
authors = ["Moru <team@moru.io>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
moru = "^0.1.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
